requirements.txt
pillow>=9.0.0
openai>=0.27.0
python-dotenv>=0.19.0
numpy>=1.19.0 